refactor: Agrupación componentes de la fase programación udistrital/sisifo_documentacion#232
