+++ src/app/app-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { GestionAuditoriaComponent } from './modules/gestion-auditoria/gestion-auditoria.component';
-import { ConsultaPlanAnualAuditoriaComponent } from './modules/consulta-plan-anual-auditoria/consulta-plan-anual-auditoria.component';
-import {RevisionPlanAuditoriaComponent}from "./modules/revision-plan-auditoria/revision-plan-auditoria.component"
-import {AsignacionAuditoresComponent}from "./modules/asignacion-auditores/asignacion-auditores.component"
-import { AuditoriasEspecialesComponent } from './modules/auditorias-especiales/auditorias-especiales.component';
-import { RegistrarAuditoriasComponent } from './modules/registrar-auditorias/registrar-auditorias.component';
-import { RegistrarPlanComponent } from './modules/registrar-plan/registrar-plan.component';
+import { ConsultaPlanAnualAuditoriaComponent } from './modules/programacion/consulta-plan-anual-auditoria/consulta-plan-anual-auditoria.component';
+import {revisionJefeComponent}from "./modules/programacion/revision-jefe/revision-jefe.component"
+import {AsignacionAuditoresComponent}from "./modules/programacion/asignacion-auditores/asignacion-auditores.component"
+import { AuditoriasEspecialesComponent } from './modules/programacion/auditorias-especiales/auditorias-especiales.component';
+import { RegistrarAuditoriasComponent } from './modules/programacion/registrar-auditorias/registrar-auditorias.component';
+import { RegistrarPlanComponent } from './modules/programacion/registrar-plan/registrar-plan.component';
 
 
 import { APP_BASE_HREF } from '@angular/common';
@@ -29,8 +29,8 @@ const routes: Routes = [
     component: RegistrarPlanComponent
   },
   {
-    path:"revision-plan",
-    component: RevisionPlanAuditoriaComponent
+    path:"revision-jefe",
+    component: revisionJefeComponent
   },
   {
     path:"asignacion-auditor",
