+++ src/app/app.module.ts
@@ -6,12 +6,12 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { GestionAuditoriaModule } from './modules/gestion-auditoria/gestion-auditoria.module';
-import { ConsultaPlanAnualAuditoriaModule } from './modules/consulta-plan-anual-auditoria/consulta-plan-anual-auditoria.module';
-import { RevisionPlanAuditoriaModule } from './modules/revision-plan-auditoria/revision-plan-auditoria.module';
-import { AsignacionAuditoresModule } from './modules/asignacion-auditores/asignacion-auditores.module'
-import { AuditoriasEspecialesModule } from './modules/auditorias-especiales/auditorias-especiales.module';
-import { FormularioAuditoriaEspecialModule } from './modules/formulario-auditoria-especial/formulario-auditoria-especial.module';
-import { RegistrarAuditoriasModule } from './modules/registrar-auditorias/registrar-auditorias.module';
+import { ConsultaPlanAnualAuditoriaModule } from './modules/programacion/consulta-plan-anual-auditoria/consulta-plan-anual-auditoria.module';
+import { revisionJefeModule } from './modules/programacion/revision-jefe/revision-jefe.module';
+import { AsignacionAuditoresModule } from './modules/programacion/asignacion-auditores/asignacion-auditores.module'
+import { AuditoriasEspecialesModule } from './modules/programacion/auditorias-especiales/auditorias-especiales.module';
+import { FormularioAuditoriaEspecialModule } from './modules/gestion-auditoria/formulario-auditoria-especial/formulario-auditoria-especial.module';
+import { RegistrarAuditoriasModule } from './modules/programacion/registrar-auditorias/registrar-auditorias.module';
 import { CargarArchivoModule } from './modules/cargar-archivo/cargar-archivo.module';
 import { GestionAuditoriaComponent } from './modules/gestion-auditoria/gestion-auditoria.component';
 import { FormularioDinamicoComponent } from './components/formulario-dinamico/formulario-dinamico.component';
@@ -33,7 +33,7 @@ import { MatDividerModule } from '@angular/material/divider';
 import { MatExpansionModule } from '@angular/material/expansion';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatRadioModule } from '@angular/material/radio';
-import { RegistrarPlanComponent } from './modules/registrar-plan/registrar-plan.component';
+import { RegistrarPlanComponent } from './modules/programacion/registrar-plan/registrar-plan.component';
 
 
 
@@ -52,7 +52,7 @@ import { RegistrarPlanComponent } from './modules/registrar-plan/registrar-plan.
     BrowserAnimationsModule,
     GestionAuditoriaModule,
     ConsultaPlanAnualAuditoriaModule,
-    RevisionPlanAuditoriaModule,
+    revisionJefeModule,
     AsignacionAuditoresModule,
     AuditoriasEspecialesModule,
     FormularioAuditoriaEspecialModule,
