+++ src/app/modules/programacion/auditorias-especiales/auditorias-especiales.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
 import { MatTableDataSource } from '@angular/material/table';
-import { FormularioAuditoriaEspecialComponent } from '../formulario-auditoria-especial/formulario-auditoria-especial.component';
+import { FormularioAuditoriaEspecialComponent } from '../../gestion-auditoria/formulario-auditoria-especial/formulario-auditoria-especial.component';
 import { MatDialog } from '@angular/material/dialog';
