+++ src/app/modules/programacion/registrar-plan/registrar-plan.component.css
