+++ src/app/modules/programacion/revision-jefe/modal-general/modal-general.ts
