+++ src/app/modules/programacion/revision-jefe/modal-motivos-rechazo/modal-motivos-rechazo.component.css
