--- src/app/modules/revision-plan-auditoria/revision-plan-auditoria.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { RevisionPlanAuditoriaComponent } from './revision-plan-auditoria.component';
-
-describe('RevisionPlanAuditoriaComponent', () => {
-  let component: RevisionPlanAuditoriaComponent;
-  let fixture: ComponentFixture<RevisionPlanAuditoriaComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [RevisionPlanAuditoriaComponent]
-    });
-    fixture = TestBed.createComponent(RevisionPlanAuditoriaComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
